multiple_tool_calling_react_agent/requirements.txt
langchain
python-dotenv
langchain-core
langchain-community
langchain-openai
requests
duckduckgo-search
